consolidated more common code from model classes into master_model.py
